ytd/requirements.txt
pycurl
feedparser
certifi
pytube3
tkinter